svn merge -r 13822:13876 http://svn.civicrm.org/civicrm/branches/v2.0 including CRM-2914, CRM-2917, CRM-2919, CRM-2925, CRM-2932, CRM-2933, CRM-2934 and CRM-2935 works/fixes
git-svn-id: http://svn.civicrm.org/civicrm/trunk@13877 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ CRM/Core/BAO/SchemaHandler.php
@@ -111,6 +111,7 @@ class CRM_Core_BAO_SchemaHandler
     }
 
     static function buildFieldSQL( &$params, $separator, $prefix ) {
+        $sql = '';
         $sql .= $separator;
         $sql .= str_repeat( ' ', 8 );
         $sql .= $prefix;
@@ -175,6 +176,7 @@ class CRM_Core_BAO_SchemaHandler
     }
 
     static function buildIndexSQL( &$params, $separator, $prefix ) {
+        $sql = '';
         $sql .= $separator;
         $sql .= str_repeat( ' ', 8 );
         if ( $params['unique'] ) {
@@ -243,7 +245,9 @@ class CRM_Core_BAO_SchemaHandler
                 $sql .= ", DROP PRIMARY KEY";
             }
             if ( CRM_Utils_Array::value( 'searchable', $params ) ) {
-                $sql .= ", DROP INDEX INDEX_{$params['name']}";
+                if ( $params['type'] != 'text' ) {
+                    $sql .= ", DROP INDEX INDEX_{$params['name']}";
+                }
             }
             if ( CRM_Utils_Array::value( 'fk_table_name', $params ) ) {
                 $sql .= ", DROP FOREIGN KEY FK_{$params['table_name']}_{$params['name']}";
